:recycle: refactor : 코드 리팩토링

See merge request s07-bigdata-recom-sub2/S07P22A107!150

Former-commit-id: 5fc3aa0e77d8bef32bd5d01841307b51d8cd7b47

+++ BE/src/main/java/com/team7/chaekin/domain/meeting/service/MeetingService.java
@@ -91,8 +91,11 @@ public class MeetingService {
     }
 
     @Transactional
-    public void deleteMeeting(long meetingId) {
+    public void deleteMeeting(long memberId, long meetingId) {
         Meeting meeting = getMeeting(meetingId);
+        if (!meeting.getMeetingLeader().getId().equals(memberId)) {
+            throw new CustomException(DomainErrorCode.ONLY_LEADER_CAN_DELETE_MEETING);
+        }
         meeting.delete();
     }
